gradleProfile utils
